Fixed bug in shell that allowed customers to return the wrong item
The bug allowed customers to give the number of the item and an id that didnt correspond with that item. It also allowed it to change the log and the inventory. I fixed this bug by changing input_guess input_choice and input_int.

Next, I plan on styling a bit.